Forgot to do pushes
I got into the zone of working on this and forgot that I was supposed to be pushing this in here. 
Currently I am up to the point of sorting. I have a problem in which it will not allow me to sort in any way except for typing, and I don't know why, but I am trying to debug it.